fix: stop hrToMin() wrapping past 1092 hours
hour * 60U is a 16-bit multiply on AVR with a uint16_t return. The other
conversions in Time already return uint32_t and use a UL literal.

+++ lib/common/src/common.hpp
@@ -37,8 +37,8 @@ public:
   /// @brief Converts hours to minutes.
   /// @param hour The number of hours to convert.
   /// @return The equivalent time in minutes.
-  static constexpr uint16_t hrToMin(uint16_t hour) {
-    return hour * 60U;
+  static constexpr uint32_t hrToMin(uint16_t hour) {
+    return hour * 60UL;
   }
 
   /// @brief Converts milliseconds to microseconds.
